Scores Commit
Now checks the scores of the dealer and the player in the event that no one 'busts' and identifies the winner. Scores have now been better generalized and are tracked throughout the game rather than re-computed on each check. Also catches the player if 21 exactly is hit. If the player hits 21 the game finishes the dealers hand and ends the game as though the user chose to stand.