Dealer Hand Commit
Dealer hand now goes to the card being flipped on the dealer side and resets to the deck position when done. Time is adjusted by position. Card deck has also been added and sits on the side between dealer and player cards. Background has been changed to remove dealer hand to allow it to move independently.